+++ packages/checkout/src/views/PendingTransaction.tsx
@@ -1,31 +1,54 @@
-import { Box } from '@0xsequence/design-system'
+import { Box, Spinner, Text } from '@0xsequence/design-system'
 import { useProjectAccessKey } from '@0xsequence/kit'
 import React, { useEffect } from 'react'
 
 import { fetchSardineOrderStatus } from '../api'
 import { TransactionPendingNavigation } from '../contexts'
-import { useNavigation, useCheckoutModal } from '../hooks'
+import { useNavigation, useCheckoutModal, useSardineClientToken } from '../hooks'
+import { useTokenMetadata } from '@0xsequence/kit'
 
 const POLLING_TIME = 10 * 1000
 
 export const PendingTransaction = () => {
   const nav = useNavigation()
   const { settings } = useCheckoutModal()
+
   const {
-    params: { authToken, orderId }
+    params: { creditCardCheckout }
   } = nav.navigation as TransactionPendingNavigation
   const { setNavigation } = nav
   const projectAccessKey = useProjectAccessKey()
 
+  const { data: tokensMetadata } = useTokenMetadata(creditCardCheckout.chainId, creditCardCheckout.nftAddress, [creditCardCheckout.nftId])
+  const tokenMetadata = tokensMetadata ? tokensMetadata[0] : undefined
+
   const isDev = settings?.creditCardCheckout?.isDev || false
+
+  const disableSardineClientTokenFetch = !tokenMetadata
+
+  const { data, isLoading, isError } = useSardineClientToken({
+    order: creditCardCheckout,
+    isDev,
+    projectAccessKey: projectAccessKey,
+    tokenMetadata: tokenMetadata
+  }, disableSardineClientTokenFetch)
+
+  const authToken = data?.token
+
   const url = isDev
     ? `https://crypto.sandbox.sardine.ai/?client_token=${authToken}&show_features=true`
     : `https://crypto.sardine.ai/?client_token=${authToken}&show_features=true`
 
   const pollForOrderStatus = async () => {
     try {
+      if (!data) {
+        return
+      }
+
+      const { orderId } = data
+
       console.log('Polling for transaction status')
-      const isDev = settings?.creditCardCheckout?.isDev || false
+      const isDev = creditCardCheckout?.isDev || false
 
       const pollResponse = await fetchSardineOrderStatus(orderId, isDev, projectAccessKey)
       const status = pollResponse.resp.status
@@ -77,6 +100,32 @@ export const PendingTransaction = () => {
       clearInterval(interval)
     }
   }, [])
+
+  if (isError) {
+    return (
+      <Box flexDirection="column" justifyContent="center" alignItems="center" gap="6" style={{
+        height: '500px',
+        width: '380px'
+      }}>
+        <Box>
+          <Text color="text100">An error has occurred</Text>
+        </Box>
+      </Box>
+    )
+  }
+
+  if (isLoading || !authToken) {
+    return (
+      <Box flexDirection="column" justifyContent="center" alignItems="center" gap="6" style={{
+        height: '500px',
+        width: '380px'
+      }}>
+        <Box>
+          <Spinner size="lg" />
+        </Box>
+      </Box>
+    )
+  }
 
   return (
     <Box alignItems="center" justifyContent="center" style={{ height: '620px' }}>